Fix release-job.sh version var
Do not append any additional characters to the
release version during tagging and pushing steps.
Allow the usage of the "$VERSION" variable as it is
provided by the user in releases/*.yaml file.
This will allow more flexibility of tag convention
for other projects.

In this case, if the user wants to use the prefixed
"v", they can define it in the releases/*.yaml as:
version: 'v1.0.0'

Issue: RELENG-2131
Change-Id: I71f32600d3e17ec117776be6764c2e45ee934bdd
Signed-off-by: Jessica Wagantall <jwagantall@linuxfoundation.org>