Update patch set 1
Patch Set 1:

This is not an example. Similarly to how we have info-schema in https://github.com/lfit/releng-global-jjb/blob/master/info-schema to verify info files, we need to have this release-schema local in global-jjb to verify releases. 
Take a look at Aric's comment here where he explicitly mentions that we need this file:
https://github.com/lfit/releng-global-jjb/blob/master/shell/release-job.sh#L50

Patch-set: 1