Update patch set 1
Patch Set 1:

The "install" goal triggers a Docker build. But the Sonar templates do not login at the LF registry so that part fails.  The "test" goal should be enough to compile the Java code and run unit tests, producing stats for Sonar.

Clean is not needed in a fresh clone. There are no previous build artifacts lying around.  If it really bothers you I will add it, it does no harm.

Patch-set: 1