Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

I think I'm failing to see the point in this change.

Not only is it modifying almost all of the single-quotes to double-quotes (for no reason that I can comprehend) this also moves global-jjb away from the style we've essentially enforced on all of our downstream projects too!

I can somewhat understand the indentation fixes for the yaml.

Patch-set: 2