Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (2 comments)
> 
> I think I'm failing to see the point in this change.
> 
> Not only is it modifying almost all of the single-quotes to double-quotes (for no reason that I can comprehend) this also moves global-jjb away from the style we've essentially enforced on all of our downstream projects too!
> 
> I can somewhat understand the indentation fixes for the yaml.

This is a completely automated change, see the patch that depends on this one. "prettier" is an opinionated style linter (https://prettier.io/) it's similar to "gofmt" in that it will reformat all the code for you so you don't even have to think about it. I hooked it up with pre-commit so that it will automatically reformat the code for us without us having to spend brain power thinking about it.

Patch-set: 2