Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review-1
> > 
> > (2 comments)
> > 
> > I think I'm failing to see the point in this change.
> > 
> > Not only is it modifying almost all of the single-quotes to double-quotes (for no reason that I can comprehend) this also moves global-jjb away from the style we've essentially enforced on all of our downstream projects too!
> > 
> > I can somewhat understand the indentation fixes for the yaml.
> 
> I am no yaml expert but it is not clear to me that any of the quotes (double & single) are even needed. I quit using them in Hyperledger months ago and have not seen any problems... yet. Yaml seems to just grabs all augments on line after the 'FOO:'. I find extra quotes very distracting when reading/reviewing.
> 
> I think the more consistent white space is GR8.

Single and Double quotes in YAML have no differentiation. The only time you need to quote strings in YAML is if you need to call on special syntax or to escape special characters.

As for this patch specifically, it's a completely automated patch that was generated by a tool called "prettier" see the next patch that depends on this one.

Patch-set: 2