Update patch set 3
Patch Set 3:

Thanks Thanh, yes we had the "stage-release" since we wanted to specify this is a "staging" and not an actual release. I believe the self release jobs will use "release" rather. The issue we are facing in ONAP is that they need to stage their Maven release first, release it and then run the stage release for Maven-Docker to produce their docker images. 
Hence they asked for the separate trigger. Do you think this is a valid request?

Patch-set: 3