Update patch set 5
Patch Set 5:

Ok so some history here. We've never had a release job before until now... the original "stage-release" triggers are about staging the release. I'd propose we standardize all of the staging triggers to be "stage-release" since that's how we've always used that keyword and the meaning of that keyword.

Which means we need to come up with a new keyword for the automated releases. The better keyword for release jobs in my opinion is "promote-release" which is the industry word I've seen in many places like "Nexus" and "Artifactory" to mean staging has been validated and we want to copy the artifacts to the final release repo. This is the meaning of "promote-release" in my opinion.

Patch-set: 5