Update patch set 5
Patch Set 5: Code-Review+1

I'm fine with this patch if we want to split the meaning of staging a maven build vs staging a docker build. I'd propose a possible improvement by making 2 triggers in that case though if you agree:

1) maven-stage job triggers with 'stage-release' or 'stage-maven-release'
2) docker-stage job triggers with 'stage-release' or 'stage-docker-release'

This allows us to keep a consistent 'stage-release' for all staging jobs however for projects that have both kinds of jobs and need to differentiate between the 2 they have the more specific keyword.

But that's just my opinion, feel free to disregard this if you don't agree.

Patch-set: 5
Label: Code-Review=+1