Handle multiple repos listed in staging-repo.
Handle multiple repos listed in staging-repo for ODL.

Also multiple release files in the same commit will not work.
exiting if this case is detected

Change merge scm to gerrit patchet not master
We actually want the merge job to check out the patchset.
as we are applying
"$PATCH_DIR/$PROJECT.bundle"

Issue: RELENG-2129
Change-Id: I55dc0a4022f93a6c3c0f389c04c88c38c53eff64
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>