Release jobs can't support triggering on branches.
This is my reasoning:
When we tag we are just tagging a ref, banches don't come into play.
we choose the ref with the log file.
example:
https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/odlparent-maven-stage-4.0.x/152/

the branch name is already baked into the log file
so no problem there.

At a later date when we pull master into a {branch-foo}
the release file is going to trigger
{project-name}-release-merge-{branch-foo}
but the repo as already been tagged.
the artfact has been released
the job will fail.

So we have to trigger on a single branch,
and I believe that means we only trigger on master.

Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>
Change-Id: I1e95e864bc2b49a62b3f7016cb2f1292a6687d9a