Add support for distribution_type container
Note: release-verify and merge will need to run on build-node
centos7-docker-8c-8g

Schema updated to require distribution_type
This patch adds support for distribution_type: container

moved functions to the top.
and the logic to choose the correct one.

container_release_file downloads log_dir/console.log.gz
And parses it to get a list of container name and version
as well as the ref to tag

verify pulls container and grabs the image_id
merge then tags and pushes the container

Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>
Change-Id: Ibc6cf1a22bd6c0136a94d2b434c6cd332c12b43f