Use --config-xml option for JJB archives
Thie --config-xml option will produce actual config xml files in the
format 'job_name/config.xml' similar to the jobs directory in Jenkins
master instead of creating generically named 'job_name' files.

This allows the jobs to be dropped into a Jenkins jobs directory
path and Jenkins will be able to pick it up.

This is potentially useful for the ability to quickly drop in the
config into a Jenkins test environment.

Signed-off-by: Thanh Ha <zxiiro@gmail.com>
Change-Id: I59517d905a2448e46c50d98a471e10d7ca0fb808