Allow skip tag on release jobs
Allow projects to skip the tagging of the Gerrit repo.
Set to "true" by default.

There are cases where some projects build different
artifacts within a pom.xml located in a subdirectory
of a .git repo. In this case, the project wants to
make separate releases for their different artifacts
but only tag the repo once.
Hence, allowing them to skip the tag part.

Change-Id: I7f601ee609d33d5a6fd1ee023db0d19fe2cf34c7
Signed-off-by: Jessica Wagantall <jwagantall@linuxfoundation.org>