Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> The error output suggests details were written to file /tmp/v/pre-commit/log/pre-commit-2.log but I don't know how to get that, I don't see the workspace being preserved here, am I just blind?

Looks like it's not archived today (maybe it should be). From the logs I see it's failing on the "prettier" check which is an opinionated code formatter. The good news is the fix is pretty simple.

* Install pre-commit locally (it should be in your package manager)
* Then run `pre-commit run --all-files` 

It will run the linters and then automatically make the changes for you. All you need to do is review what it changed (`git diff`) and then commit and repush the patch.

Hope this helps.

Patch-set: 4
Reviewer: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=0