Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Thanks Anil!  That helps. But that file looks kind of broken -- it lists a "release" index-server but provides no property-details area for it.  Also I tried browsing the nexus3.acumos.org/repository/PyPi.staging repo but the nexus server shows errors on both the asset browser and component browser screens:
> 
>     Path "browse/browse/assets" not found
> 
> So it's tough to know if any Python distribution artifacts have ever been pushed at merge time.

Chris, I would assume the lack of "release" details is because the expectation is to manually release from staging to release (why there would then be pypi upload credentials though is beyond me, so it's just a theory).

There are artifacts in staging, and one in release. Please try these links:
https://nexus3.acumos.org/#browse/browse:PyPi.staging
https://nexus3.acumos.org/#browse/browse:PyPi.release

Patch-set: 5
Label: Code-Review=0