Fix use-release-file varaible
Rename "RELEASE_FILE" parameter to "USE_RELEASE_FILE"
in release-jobs.
This will match the actual varaible default value
better and will not collide with the local "release_file"
in the script.

Fix "USE_RELEASE_FILE" if statement. We are now using
a bool instead of a string.
Changing the if statements to evaluate bools.

Change-Id: I9909a16487fabacc0261f6489ac81a00cbf0f965
Signed-off-by: Jessica Wagantall <jwagantall@linuxfoundation.org>