Update patch set 5
Patch Set 5: Code-Review-2

git-review is a version from 2014 that is coming from the ansible role.
It should be removed from the ansible role. we should not be installing system level python packages.

Also. argparse was removed as a requiement of git-review in 2014.

Also this is a bad way to fix an error. just throwing pip installs in the first place you look.

Please try to find the root of the issue (rather than arguing with me and making me do it)
we have enough technical debt with all this python pinning etc already.

Patch-set: 5
Reviewer: Gerrit User 12 <12@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=-2