Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > No one addressed my comments :( .. why is the need of having this
>  > if the dry run does exactly what the verify does already... As i
>  > mentioned before, it would be nice if dry run at least displays the
>  > echos of the commands or something more than what the verify
>  > does...
>  > 
>  > well.. i hope it was at least tested :(
> 
> Jessica, a DRY_RUN is for merge job _only_ for those projects that dont want to include .release/z.x.c.yaml files in the repos or use gerrit triggers only useful for the verify jobs when enabled. 
> 
> For ODL we are not going to have the release verify jobs. The DRY_RUN is is useful to confirm that nothing is broken between the reops, so that we don't end up with releasing one of the nexus while the other has errors, something that happened recently. :(
> 
> I'm not sure what should go into the release notes, since "build with parameters" option was not introduced in this change. If you look at the verify jobs they still have the "build with params" which never gets used. If you are suggesting that needs to be removed then i can submit a change for fixing that?

Thanks for replying Anil,
I was suggesting adding in the release notes that the dry run feature is mainly developed for people who run "build with params" manually. Because it was not clear (or at least to me) that this feature was specifically for this case. Even though we are not making changes in the build with params feature, I think it would be nice to say that that this dry run feature is for people who use them.

Patch-set: 5