Update patch set 1
Patch Set 1:

this should also document the two "version" key-value pairs in the container release yaml file more clearly. As Jessica explained to me:

The version in #4 should be the version we will releases the images with. In this case 3.0.0 is what will be released as
The one in line #7 should be what is expected to be pulled. In ONAP's case for example would be "3.0.0-20190903123456Z"

Patch-set: 1