Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

Please add a/the Jira issue ID to the commit message.

You mentioned in a comment that the macro must be configured with the Nexus registry (host:port), but the doc/RST file doesn't mention how to do that.  (Or am I blind again?)

Patch-set: 4
Label: Code-Review=-1