Update patch set 2
Patch Set 2:

Ready for your review. This updates the doc for the existing code, addresses the need for Centos, Sigul and Docker; also explains the multiple version strings in a container release file. But it will need to be updated if Jessica's change to a key from version to container_image_tag is merged first.

Patch-set: 2
Label: Code-Review=0