Update patch set 1
Patch Set 1:

This removes the passive voice, silencing the completely infuriating coala "WriteGood" bear warning, but the imperative voice used here makes it sounds like directions are being given to the person using template!  "Pull the image".   Instead how about describing the actions of the template like a Javadoc-style annotation on a method; e.g., "Pulls the container with the specified version and tags with with the specified "container_tag_name" ?

Patch-set: 1