Update patch set 2
Patch Set 2:

Why not just PYTHON instead of PYTHON_VERSION. The use of the 'python-version' macro inside the yaml makes sense. The way it is used in the script makes it pretty obvious what is going on. I have seen PYTHON used in scripts before. Not going to fight for this, just a thought.

The use of the python-version that can effect multiple jobs is a trade-off. Requiring PYTHON be set for each job can be a pain, but moving to a new version of python can be hard when changing it in one place effects a large number of jobs that all need to be tested at once. This is particularly true of YAML as it can be hard to know all the jobs that might be effected by a change to a shared macro/template.

Patch-set: 2
Reviewer: Gerrit User 182 <182@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Verified=0