Update patch set 1
Patch Set 1:

(3 comments)

Part of reason I have struggled with the -l is that in almost 35+ years with Unix I have never seen this done. I knew was not a good, just could not articulate why. After a great deal of thought, hopefully I am able to articulate the problems.

Patch-set: 1