Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

I think that the removal of the --quiet on the pip install isn't going to go over very well. We've been adding it in places where it wasn't already in place because of the amount of noise.

I think we need a job level variable to enable / disable, that defaults to keeping --quiet, if you want to make it noisier.

Patch-set: 3
Reviewer: Gerrit User 1 <1@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=-1