Update patch set 7
Patch Set 7:

Thanks Anil, I did commented on your approach. 
I mentioned we should do this in a new change since the purpose of this one was just to move code and re-organize it but not refactoring it yet. 
Let's start a new change with the refactoring
I like doing small patches to make them easy to test/review

Patch-set: 7
Label: Code-Review=0