Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Below is the complete extra output from python -m pip install tox-pyenv WITHOUT the --quiet flag.  If these 20 extra lines are TRULY going to push the LF log servers into quivering immobility then fine, I will yield but I think amidst the ten of thousands of output lines this is immaterial.
> 
> --
> 
> Collecting tox-pyenv
>   Downloading 

--[snip]--

I was mostly pushing back because we've had several projects push on us to _reduce_ the log chatter from things like pip. I would rather that we make it tunable, than an either / or because I don't want to be in the constant middle of adding / removing these things because one project hates all the extra spew and another wants it!

Patch-set: 3