Update patch set 4
Patch Set 4:

> What nit?  The "--quiet" option is in the shell script. I restored
 > to address AG's comment.
 > 
 > I hear people asking me to add methods for selecting silent/verbose
 > options in this change and/or write logs to a special place.  I
 > would like to defer all of that, unless there is well-established
 > practice in other global-jjb templates that I could copy.

I was referring to older version of the patchset which did not show ---quiet was restored. So I recommended the --log option that is supported by `pip install`, since we'd like to put the minimal info console logs and logs specific tools/utils can go into spearate folders/files in $WORKSPACE/archives. We are doing this for tox, packer jobs, syslogs/sar reports etc. but there is always more room for improving things.

Patch-set: 4