Update patch set 5
Patch Set 5:

> Patch Set 3:
> 
> I'm ok with moving all scripts to this approach. However, this should be tested with care. 
> 
> I'm pretty sure we were using "$PYHTON -m pip ..." and rolled back this as a workaround, am unable to recall the exact reason why? Thanh would know.
> 
> https://git.opendaylight.org/gerrit/#/c/releng/builder/+/58782/

I'm not sure what else to add here. The link above explains why `$PYTHON -m` was originally needed in the body of the commit message pretty clearly but let me know if there's a specific part of the explanation that's confusing and I can try to clarify.

Patch-set: 5