Update patch set 16
Patch Set 16: Code-Review-1

Thanks Eric but this is NOT ready for merge.  I should maybe make this work-in-progress to indicate that clearly?  I'll post -1 for now.

The challenge is guarding the release action, the publish-to-pypi step.  That should only be done by committers, not by any old drive-by gerrit-comment writer.  I think there should also be some minor discussion of what staging means here, and whether it's acceptable to build a release version from scratch, instead of copying it from a known place.

Patch-set: 16
Label: Code-Review=-1