Add template to update OpenStack cloud images
This job finds and updates OpenStack cloud images on the
ci-management source repository.

The job is triggered in two ways:

1. When a packer merge job completes, the new image name created is
   passed down to the job.
2. Manually trigger the job to update all images.

When the job is triggered through an upstream packer merge job, this
only generates a change request for the new image built.

When the job is triggered manually, this job finds the latest
images on OpenStack cloud and compares them with the images currently
used in the source ci-management source repository. If the compared
images have newer time stamps are **all** updated through a change
request.

This job requires a Jenkins configuration merge and verify job setup and
working on Jenkins.

Issue: RELENG-2352
Change-Id: I2c75407bbe9d6866ded492310b8d34b47c3a1341
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>