lf-infra-pre-build pyenv
This allows pyenv to be set
by adding
pyenv local 3.6.9
To any script that runs subsequent to this one.

going forward, we will move away from bash -l
and have our own lfenv.sh script that we source.

But for now it provides an easy way of choosing a
python version

Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>
Change-Id: I4fed30b973ec2a086436cade223e43c7e060a062