Update patch set 6
Patch Set 6:

(1 comment)

I think we need to have a meeting to talk about this..... 

I am not an expert with pyenv/virtualenv but I do know that its common interactive usage is quite different that how it should be used from within a script.

Using a command from a virtual env does not require ANY changes to the  environment. Note how I run jenkins-jobs locally: 'alias jj=/home/tijohnson/.local/bin/jenkins-jobs'. No variables to set, no changes to the PATH, no need to source 'activate' script (it assumes the pyenv directory already exists). I run 'openstack' exactly the same way. It works this way by design to conform with standard Unix practices. All hell would break loose if it did not.

Also this script (python-tools-install.sh), is already pretty messed up because it is optional. It may be called before build, after build or both (I might be mistaken here). Probably not the right place to put these changes. Not sure where the right place is....

Putting this in .bash_profile may not work with '-l' on all distros.

Even if we did put this in a file that is sourced, rule #1 for sourcing is that it should be treated as a library. Set shell/env vars, create aliases/functions but do not 'execute' code.
This is a rule, not a prime directive, but is rarely needs to be broken.

Patch-set: 6