Update patch set 6
Patch Set 6: Code-Review-1

If unset variable is a acceptable you should handle that case. We always want the default to be Error because of unset variables we don't expect. Even if it 'works' when script is created, things may change in the future.

The only reason that -e is not the default is because by default bash is run as interactive shell.

Patch-set: 6
Reviewer: Gerrit User 182 <182@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=-1