Update patch set 22
Patch Set 22: Code-Review-1

(4 comments)

We need step back from this and revisit the decision to use pyenv. First of all it doubles down the '-l' debacle. I have no problem with pyenv in interactive shell. But it is a bad choice for complex batch scripting. It requires sourcing third party scripts that we do not control or probably fully understand.

I suspect that if shellcheck were to check the functions created by pyenv it would not be happy, but the '-l' gets around that. Its one thing adding this to a project, but adding this to global-jjb is another thing altogether.

Patch-set: 22
Label: Code-Review=-1