Update patch set 22
Patch Set 22:

To give the perspective from a python developer who contributes python code to all 3 ATT/LF projects (onap, oran, acumos):

from the python dev perspective, we don't need and don't care that pyenv is installed; that is an abstracted detail. What we do care however, is that all simultaneous python versions that our tox depends on is installed in the environment that the verify job runs in. That is, if my tox tests py27,py34,py35,py36,py37, then all of those python versions need to be on the base system for tox to succeed. And this is common, to run unit tests in numerous python versions to catch errors that may be present in some user's envs because of python changes. And across all python projects you don't know what python versions each repo depends on. 

So, I treat pyenv as an implementation detail of the underlying CICD system. Equally fine would just be base images with all known python versions installed (27 through 38) so that tox jobs succeeds.  I'm guessing the reason pyenv was used is that pyenv makes that managable; this is the exact problem pyenv is trying to solve, simultaneously installing python versions easily.

Patch-set: 22