Update patch set 22
Patch Set 22:

Agreed, pyenv provides some great functionality and does an excellent job of 'abstracting' the user from the gory details. Once people start using it, it will be difficult to take it away. And that is why I hope this never gets merged.

Unfortunately we (LF) need to fully understand/control how python is used/versioned. And that process needs to be reviewable. The python version/environment is NOT that complicated.

Part of the problem I am having is that I don't have a detailed list of requirements for python versioning. That comment from Tommy Carpenter was very enlightening.

Patch-set: 22