Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> The code looks good, but would it be better if we just add a conditional step rather than a copy of the templates?
> This way, people will not need to migrate to a different job but just add a variable called "run-sonar-prescan: true" for example.

After doing some testing, I'm not sure a conditional will work. The issue is that the builder needs to be defined by the caller, but even if the condition is false, jjb expects the builder that it refers to to be present. So if I have:

    builders:
      - conditional-step:
          condition-kind: boolean-expression
          condition-expression: "{prescan}"
          steps:
            - lf-sonar-prescan # Must be defined by caller

JJB errors out because lf-sonar-prescan doesn't exist. I think it might actually require its own template in order to avoid this. But if there's a way to simplify it that I'm missing, please let me know.

Patch-set: 2
Reviewer: Gerrit User 192 <192@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=0