Update patch set 34
Patch Set 34:

I will continue to fiddle the find-files bits, but my testing shows that that ls does the right thing and emits each file on a separate line when used in a pipeline. It only outputs columns when the output is a TTY. And in the (awful) case of a file having embedded spaces, wc -w answers a different result from wc -l.

    (venv-py36) mbp-cl778h:~/git/onap/ci-management/global-jjb$ ls
    README.md				releasenotes/
    archives/				relnotes/
    check_jjb_version.py			requirements.txt
    docs/					schema/
    ensure-documented.sh*			shell/
    jenkins-admin/				tox.ini
    jenkins-init-scripts/			wss-unified-agent.config.example
    jjb/					yaml-verify-schema.py
    jjb-compare-xml.sh*			yamllint.conf
    license-header.txt
    (venv-py36) mbp-cl778h:~/git/onap/ci-management/global-jjb$ ls | wc -l
      19

Patch-set: 34
Label: Code-Review=0