Update patch set 7
Patch Set 7:

What are the prereqs here in minions to use this?  I wonder if possibly this change is the reason for the latest breakage in Acumos: lf-tools-install.sh fails.  It's very possible our build minions just are not ready but boy it's tough to know.

Here's a link to a failed job.  The problem also means the logs don't get archived:

https://jenkins.acumos.org/job/federation-maven-dl-verify-master-mvn35-openjdk11/28/console

I opened a ticket:

https://jira.linuxfoundation.org/servicedesk/customer/portal/2/IT-17714

Patch-set: 7
Reviewer: Gerrit User 114 <114@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Verified=0