Update patch set 1
Patch Set 1:

> Please help me understand, what happens after sigul signs an
 > artifact?  Who (or what) benefits? Does PyPI recognize this sort of
 > signed artifact, and do something better/faster/more secure?

The only way consumers can verify the released binaries downloaded, are exactly the ones that were released is by checking the cryptographically signatures, and making sure this is signed by project maintainers keys (or using a trusted third party like the sigul that does the singing on behalf of the maintainer). Pypi does not care or require these signatures, and same goes with Nexus. However, these are release best practices followed in all projects.

Patch-set: 1