Update patch set 1
Patch Set 1:

Thanks for the details and link to shell script.  I can extend PyPI templates to call it.  Please guide me, what exactly should be signed?  Probably you want this script to generate a signature for all source and binary distributions created by setup.py in the dist/ folder, is that right?  I guess that signature is saved in the Jenkins build log(s) and that's the goal?

Patch-set: 1