Update patch set 26
Patch Set 26:

(3 comments)

> Patch Set 23: Code-Review-1
> 
> (31 comments)
> 
> Please consider my comments.  I'm voting -1 because I've never seen a verify job that is permitted to make changes anywhere other than in the build minion, and I don't think this should stray from that precedent.


Hi, Since there is an unknown amount of time between when a project is created and when we can trigger a job VIA the api, I decided to create to project during the verify job.

normally this will be a no-op as the project will already exist, but eliminate the need for intervention when a new project starts making docs for the first time.

Patch-set: 26