Update patch set 1
Patch Set 1: Code-Review-1

if git tag -v "$VERSION"; then
          echo "ERROR: Repo already tagged"
          exit 1
      else

Okay so, if someone pulls master into {branch} that will bring in the release file
and trigger this job, we will then get a -1 on the change.

we should exit 0 here so that the job exits cleanly.
https://gerrit.linuxfoundation.org/infra/c/releng/global-jjb/+/16441

In general I think that a tag existing should be considered a success, the work is already done.

Patch-set: 1