Update patch set 1
Patch Set 1:

Thanks for the analysis of the tagging behavior in the shell script but that's not what I'm changing here.  Is it usual practice to add comments about other unrelated problems on a newly opened change set on the same file?

I don't really care one way or the other, but I modeled this after the release-job.sh where I believe an existing tag is treated as a fatal error -- I was just trying to be consistent with established LF practice.

Patch-set: 1