Update patch set 1
Patch Set 1:

> Okay so, if someone pulls master into {branch} that will bring in the release file
> and trigger this job, we will then get a -1 on the change.

I'm not sure what you mean exactly by "pulls master into {branch}". In my experience, if someone rebases a branch onto master, the release file becomes available in that branch but it will not appear in the list of changed files, so the release job will not be triggered.

Patch-set: 1