Update patch set 4
Patch Set 4:

(6 comments)

You have not included the yaml/config changes that would add 'utility-venv.sh' to the build.

Since this is a Python venv specifically created for pypi operations I think the old location is really preferable or maybe /tmp/pypi-venv? '~' is not really a good place to create temporary files/dirs.

I have created a function 'lf-activate()' that validates the existence of the venv and then adds its 'bin' to the PATH. 'lf-activate()' is defined in a script that you can source:

source ~/lf-env.sh
lf-activate $venv

'lf-activate()' is available in global-jjb >= v0.45.0

Patch-set: 4