Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I didn't realize you intended this VENV exclusively for PyPI scripts, is that truly necessary? It seems reasonable, in the interest of accelerating builds by reducing unnecessary pip installs, to share this venv among any scripts that require python packages perhaps including lf-install. Maybe I don't understand a blocking issue here tho? Please consider it, thanks.

I would agree and I have been looking at moving global-jjb in that direction. We would want the 'master' VENV to be created from a single place. It would have to be based on the 'version' matrix (see python-tools-install.sh) that are required for all the tools. For just lftools it is pretty simple, for tox, lftools, niet..... it is pretty complicated.  And of course when the version matrix is in-correct you get a nice error message from 'pip' but it still returns a zero status!

Patch-set: 4