Update patch set 1
Patch Set 1:

I see you changed from upper to lower case. It seems to me this the only point of this macro is to set an environment variable for shell scripts that follow invocation of this script. But if you change the case, that doesn't work anymore.  A quick search reveals that MAVEN_OPTIONS is used in these scripts:

./shell/maven-sonar.sh
./shell/maven-build-deps.sh
./shell/deploy-maven-file.sh
./shell/maven-javadoc-generate.sh
./shell/maven-build.sh

Of course I don't know if any of these rely on lf-env.sh being run first, or if instead they actually rely on script ./shell/common-variables.sh being run first, which also defines this same environment variable.

Is this lf-env.sh script intended to replace common-variables?  I just don't understand the intent of this change.

Patch-set: 1