Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What I'm trying to understand is, what other macros, scripts, templates, what have you depend on MAVEN_OPTIONS?  You didn't change anything else, just stopped setting this upper-case shell environment variable. I see that referenced in:

It is just a 'library' (it does not 'do' anything, it just defines functions) and at this time no-one is using it. The 'MAVEN_OPTIONS' variable is currently being used in many places and it is initialized using a different process. Once the 'library' is in place, I can start changing the way the variable is set. One of the 'problems' with this library is you have to wait for a release (tag) of global-jjb before you see the change to the library in a Jenkins build.

The fact that no-one is using it is why I justify no release note for this change.

Patch-set: 1