Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > I didn't realize you intended this VENV exclusively for PyPI scripts, is that truly necessary? It seems reasonable, in the interest of accelerating builds by reducing unnecessary pip installs, to share this venv among any scripts that require python packages perhaps including lf-install. Maybe I don't understand a blocking issue here tho? Please consider it, thanks.
> 
> I would agree and I have been looking at moving global-jjb in that direction. We would want the 'master' VENV to be created from a single place. It would have to be based on the 'version' matrix (see python-tools-install.sh) that are required for all the tools. For just lftools it is pretty simple, for tox, lftools, niet..... it is pretty complicated.  And of course when the version matrix is in-correct you get a nice error message from 'pip' but it still returns a zero status!

What exactly is the purpose of the python-tools-install.sh script? I see it does a before and after diff of what may be installed, and generates a requirements.txt file, but I'm not sure why?

(IMHO at the least the requirements list should be broken off into an independent requirements.txt file if possible)

And what does "pretty complicated" here mean? If we have specific version requirements, we identify those in the requirements.txt, right? What else is under the hood that adds to the complexity here?

Also, can you give an example error condition for pip where it exits zero still? I want to research why it's doing that and find a manageable solution. Thanks.

Patch-set: 4